Refactor block groups away from gpio.
