+++ src/modules/editor.js
@@ -35,9 +35,11 @@ import './blockly/generators/propc/communicate';
 import './blockly/generators/propc/control';
 import './blockly/generators/propc/cogs';
 import './blockly/generators/propc/custom_code';
+import './blockly/generators/propc/eeprom';
 import './blockly/generators/propc/gpio';
 import './blockly/generators/propc/oled';
 import './blockly/generators/propc/heb';
+import './blockly/generators/propc/pins';
 import './blockly/generators/propc/procedures';
 import './blockly/generators/propc/s3';
 import './blockly/generators/propc/sd_card';
